Move FormMetadata into app_schemas
